agents/the-amplifier/pyproject.toml
[tool.poetry]
name = "the-amplifier"
version = "0.1.0"
description = "ZeroClick AI-native ad integration — monetize agent attention with contextual ads"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
payments-py = {version = ">=1.1.0", extras = ["mcp"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
